requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8
numpy>=1.24
fastapi>=0.110
uvicorn>=0.27
pillow>=10
requests>=2.31
pydantic>=2.0